+++ assets/apps/dashboard/src/Components/Content/FreePro.js
@@ -4,12 +4,8 @@ import FeatureRow from '../FeatureRow';
 import { __ } from '@wordpress/i18n';
 import { Button } from '@wordpress/components';
 
-const Pro = ({ features, url }) => {
-	if (typeof neveDash !== 'undefined') {
-		const { featureData, upgradeURL } = neveDash;
-		features = featureData;
-		url = upgradeURL;
-	}
+const Pro = () => {
+	const { featureData } = neveDash;
 	return (
 		<div className="col">
 			<table className="card table">
@@ -19,7 +15,7 @@ const Pro = ({ features, url }) => {
 						<th className="indicator">Neve</th>
 						<th className="indicator">Neve Pro</th>
 					</tr>
-					{features.map((item, index) => (
+					{featureData.map((item, index) => (
 						<FeatureRow key={index} item={item} />
 					))}
 				</tbody>
@@ -35,7 +31,7 @@ const Pro = ({ features, url }) => {
 				<Button
 					target="_blank"
 					rel="external noreferrer noopener"
-					href={url}
+					href={neveDash.upgradeURL}
 					isPrimary
 				>
 					{__('Get Neve Pro Now', 'neve')}
